model-training-service/src/two_tower_training/requirements.txt
torch==2.0.1+cpu
pandas
pyarrow
pyyaml
mlflow
scikit-learn
kfp
python-box
